requirements.txt
javabridge==1.0.18
python-bioformats==1.5.2
numpy==1.15.4
joblib==0.13.0
tqdm==4.28.1
h5py==2.9.0
